Country/State field fix
The Billing and Shipping Country Fields were using the abbreviation for the country. The solution is either to use the 2 digit country code(not the 3 digit USA) or the full name. I updated the fields to use the full name for the Country and state.
Reordered the fields so that the country is checked first, so that the field dependency doesn't cause an issue. 
Class now passes validation checks